refactor(App): streamline AppLayout and routing logic
- Removed unused imports and components from AppLayout.
- Simplified navigation structure by integrating cost and IAM routes.
- Enhanced sidebar toggle functionality.
- Updated App component to include IamProvider for context management.
- Improved routing logic to dynamically set the basename based on IAM path.

This refactor aims to improve code clarity and maintainability while ensuring proper context handling for IAM and cost management routes.


feat(App): implement chrome stub and routing configuration

- Removed the on-prem remote constants file as it is no longer needed.
- Introduced a chrome stub for authentication and user management, simulating the insights environment.
- Added routing configuration for cost management and IAM, defining paths and titles for navigation.
- Established a Scalprum configuration to manage app metadata and module loading.

These changes enhance the application's structure and prepare it for improved integration with the insights environment.


refactor(App): centralise scope constants and move remoteKey into Router

Move costScope and rbacScope from Router into scalprum-config alongside
the config entries they key into. Make Scope and ScalprumConfig
module-private since no consumer imports them. Router now derives
remoteKey internally from useIam(), so the prop is dropped and AppLayout
calls <Router /> with no arguments.


refactor(koku-ui-onprem): flatten App sub-components into src top-level dirs

Move components, routes, scalprum config, chrome stub, and API helpers out
of the nested App/ subdirectory into focused top-level directories:
- src/components/   — App, AppLayout, AppRoutes, AppToolbar, IamContext,
                      NavItem, UserMenu
- src/data/         — api, routes, scalprum
- src/stubs/        — chrome

Update tsconfig.json with the '#/*' path alias and bootstrap.tsx imports.


fix(dev): fall back to x-auth-request-user in /api/me when preferred_username absent

oauth2-proxy only injects X-Auth-Request-Preferred-Username when the OIDC
token contains a non-empty preferred_username claim.  When that claim is
missing (e.g. older proxy image or Keycloak client without profile scope
mapper), the previous handler logged a noisy per-request warning and
returned 'dev-user' as the username.

Fall back to x-auth-request-user (set from sub/email claim, always present)
and x-auth-request-email for the email field.  Only emit the warning when
*no* oauth2-proxy auth headers are present at all, which indicates a true
proxy bypass rather than a missing optional claim.
