+++ apps/koku-ui-onprem/src/data/api.ts
@@ -1,0 +1,11 @@
+import { createAxiosInstance } from '@koku-ui/ui-lib/api';
+
+export interface UserData {
+  username?: string;
+  email?: string;
+}
+
+export const fetchCurrentUser = (): Promise<UserData> =>
+  createAxiosInstance()
+    .get<UserData>('/api/me')
+    .then(res => res.data);
