+++ apps/koku-ui-onprem/tsconfig.json
@@ -4,6 +4,7 @@
     "baseUrl": "./src",
     "outDir": "./dist/",
     "paths": {
+      "#/*": ["./*"],
       "@koku-ui/ui-lib/*": ["../../../libs/ui-lib/src/*"],
       "@koku-ui/onprem-cloud-deps/*": ["../../../libs/onprem-cloud-deps/src/*"],
     }
